Moved listener factories into `ZF\Apigility\Admin\Listener` namespace
